refactor(monument): consolidate command and query services into a unified service
- Merge `MonumentCommandService` and `MonumentQueryService` into a single `MonumentService`
- Consolidate `create` and `update` methods into a single `save` method within the new service
- Update `MonumentController` and `MonumentAdminController` to use the unified `MonumentService`
- Refactor `MatchReviewProcessor` and `DiscoveryReviewProcessor` to interact directly with `MonumentRepository` instead of service classes
