+++ monument/src/main/java/pt/estga/monument/services/MonumentService.java
@@ -6,9 +6,10 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pt.estga.monument.Monument;
+import pt.estga.monument.MonumentMapper;
 import pt.estga.monument.MonumentRepository;
 import pt.estga.monument.dots.MonumentListDto;
-import pt.estga.monument.MonumentMapper;
+import pt.estga.sharedweb.exceptions.ResourceNotFoundException;
 import pt.estga.sharedweb.filtering.QueryProcessor;
 import pt.estga.sharedweb.models.PagedRequest;
 import pt.estga.sharedweb.models.QueryResult;
@@ -18,11 +19,11 @@ import java.util.Optional;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class MonumentQueryService {
+public class MonumentService {
 
     private final MonumentRepository repository;
-    private final QueryProcessor<Monument> queryProcessor;
     private final MonumentMapper mapper;
+    private final QueryProcessor<Monument> queryProcessor;
 
     public Optional<Monument> findById(Long id) {
         return repository.findById(id);
@@ -45,5 +46,18 @@ public class MonumentQueryService {
 
     public Page<Monument> findByDivisionId(Long id, Pageable pageable) {
         return repository.findByDivisionId(id, pageable);
+    }
+
+    @Transactional
+    public Monument save(Monument monument) {
+        return repository.save(monument);
+    }
+
+    @Transactional
+    public void deleteById(Long id) {
+        Monument existing = repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Monument with id " + id + " not found"));
+
+        repository.softDelete(existing);
     }
 }
