+++ review/src/main/java/pt/estga/review/processors/MatchReviewProcessor.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Component;
 import pt.estga.mark.entities.Mark;
 import pt.estga.mark.repositories.MarkRepository;
 import pt.estga.monument.Monument;
-import pt.estga.monument.services.MonumentQueryService;
+import pt.estga.monument.MonumentRepository;
 import pt.estga.review.dtos.DiscoveryContext;
 import pt.estga.review.enums.ReviewType;
 import pt.estga.review.models.ResolutionResult;
@@ -18,7 +18,7 @@ import pt.estga.review.models.ResolutionResult;
 public class MatchReviewProcessor implements ReviewProcessor {
 
     private final MarkRepository markRepository;
-    private final MonumentQueryService monumentQueryService;
+    private final MonumentRepository monumentRepository;
 
     @Override
     public ReviewType getSupportedType() { return ReviewType.MATCH; }
@@ -32,7 +32,7 @@ public class MatchReviewProcessor implements ReviewProcessor {
             mark = markRepository.findById(context.existingMarkId()).orElseThrow();
         }
         if (context.existingMonumentId() != null) {
-            monument = monumentQueryService.findById(context.existingMonumentId()).orElseThrow();
+            monument = monumentRepository.findById(context.existingMonumentId()).orElseThrow();
         }
 
         return new ResolutionResult(mark, monument);
